backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
scikit-learn
opencv-python-headless
numpy
requests
pydantic